Gallery and InRibbonGallery filters support for MVVM, MenuItem and Toolbar support for MVVM.
Also fixed name for IsRibbonAlignedProperty of ScreenTip (Registered as
BelowRibbon?) to be able to bind to it. Added EnumToEnumConverter to
support binding of enums like Size. Added SizeDefinitionValueConverter
for binding SizeDifinitions. Added
RibbonToolBarControlDefinitionConverter,
RibbonToolBarControlRowChildDefinitionConverter,
RibbonToolBarRowConverter and RibbonToolBarLayoutDefinitionConverter to
add support for toolbar layout in MVVM case.
Fixed Office 2010 combobox non editable template. (Mouse over a combobox
item will not highlight the item).
Fixed Office 2010 control template for toolbar. The separator before a
group of controls is not displayed.